% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_occ_expert.R
\name{check_occ_expert}
\alias{check_occ_expert}
\title{Check which occurrence points fall outside of known range}
\usage{
check_occ_expert(range, occ_data)
}
\arguments{
\item{range}{polygon defining the species range (can be an \code{sf} or \code{sp} object)}

\item{occ_data}{an \code{sf} spatial data frame with occurrence points}
}
\value{
The \code{occ_data} input with four columns added (core:buff3)
}
\description{
The known range can be either a IUCN range, hand-delineated expert polygon or QDS cells.
}
\details{
Explain what the function does
}
\note{
It's important to note a couple of things. The buffers are set at 5, 10, and 20 km
}
\examples{
\dontrun{
data <- check_occ_expert(range, occ_data)
}
}
\author{
Dominic Henry
}
