% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_occ_range.R
\name{check_occ_range}
\alias{check_occ_range}
\title{Check which occurrence points fall outside of known range}
\usage{
check_occ_range(range, occ_data)
}
\arguments{
\item{range}{polygon defining the species range (can be an \code{sf} or \code{sp} object)}

\item{occ_data}{an \code{sf} spatial data frame with occurrence points}
}
\value{
The \code{occ_data} input with four columns added (\code{core}, \code{buff1}, \code{buff2}, and \code{buff3})
}
\description{
The known range can be either a IUCN range, hand-delineated expert polygon or a QDS shapefile.
}
\details{
The function creates three buffers around the range and identifies where a point is located based on the buffers (i.e., within the core range, or within a 5km, 10km, or 20km buffer). Four columns are added to the input data frame (\code{core}, \code{buff1}, \code{buff2}, and \code{buff3}) and are given a logical value depending on where the point is located (\code{TRUE} if the point is in the core or buffers, \code{FALSE} if it isn't).
}
\note{
If a point is located beyond the 20km buffer then all of \code{core}, \code{buff1}, \code{buff2}, and \code{buff3} will be given the value \code{FALSE}.
}
\examples{
\dontrun{
data <- check_occ_range(range, occ_data)
}
}
\author{
Dominic Henry
}
