% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_occ_temporal.R
\name{plot_occ_temporal}
\alias{plot_occ_temporal}
\title{Create temporal plots of species occurrence data}
\usage{
plot_occ_temporal(data, ptheme)
}
\arguments{
\item{data}{Simple Feature point object with species occurrence data}

\item{ptheme}{ggplot theme}
}
\value{
Temporal JPEG plots
}
\description{
This function summarises occurrence data in a temporal context.
}
\details{
The function draws two plots  showing occurrence record frequencies and writes them to the species directory.
}
\note{
The size of the text and other elements of the plots can be change by customising the \code{theme} object.
}
\examples{
\dontrun{
plot_occ_temporal(occ_data_sf, ptheme)
}
}
\author{
Dominic Henry
}
