% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMutils-package.R
\docType{package}
\name{SDMutils-package}
\alias{SDMutils}
\alias{SDMutils-package}
\title{SDMutils: Helper functions for an SDM workflow}
\description{
This package provides a set of commonly used functions that are used to summarise occurrence records in both space and time, identify outliers, prepare data for SDM input, and write output files from the various SDM steps in a standardised format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DomHenry/SDMutils}
  \item Report bugs at \url{https://github.com/DomHenry/SDMutils/issues}
}

}
\keyword{internal}
