% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bart_plots.R
\name{write_bart_plots}
\alias{write_bart_plots}
\title{Generate maps of BART SDM predictions}
\usage{
write_bart_plots(data, mean_theme, ci_theme)
}
\arguments{
\item{data}{a \code{RasterStack} object returned by \code{embarcadero::predict2.bart()}}

\item{mean_theme}{a \code{rasterTheme} used for plotting the mean probability SDM map}

\item{ci_theme}{a \code{rasterTheme} used for plotting the posterior width (95\% credible interval) SDM map}
}
\value{
The following plot files:
\enumerate{
\item \code{SDM_mean_probability.pdf}
\item \code{SDM_mean_probability.png}
\item \code{SDM_posterior_width.pdf}
\item \code{SDM_posterior_width.png}
\item \code{SDM_four_panel_probability.pdf}
\item \code{SDM_four_panel_binary.pdf}
\item \code{SDM_mean_binary.pdf}
\item \code{SDM_mean_binary.png}

}
}
\description{
This function writes plots using a combination of PDF and PNG outputs.
}
\details{
The function writes 8 plots related to the species SDM predictions, both probability and binary surfaces. The species' occurrence points have been overlaid on certain plots. Plots are written to the species directory.
}
\examples{
\dontrun{
write_bart_plots(data = map, mean_theme = maptheme_mean, ci_theme = maptheme_ci)
}
}
\author{
Dominic Henry
}
