% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bart_rasters.R
\name{write_bart_rasters}
\alias{write_bart_rasters}
\title{Write BART raster predictions to GeoTIFF format}
\usage{
write_bart_rasters(data)
}
\arguments{
\item{data}{a \code{RasterStack} object returned by \code{embarcadero::predict2.bart()}}
}
\value{
Seven raster layers:
\enumerate{
\item \verb{SDM_prob_mean_\{sppselect\}}: Mean probability of habitat suitability
\item \verb{SDM_prob_lower_\{sppselect\}}: Lower 95\% credible interval probability of habitat suitability
\item \verb{SDM_prob_upper_\{sppselect\}}: Upper 95\% credible interval probability of habitat suitability
\item \verb{SDM_bin_mean_\{sppselect\}}: Mean binary presence/absence prediction (based on TSS threshold)
\item \verb{SDM_bin_lower_\{sppselect\}}: Lower 95\% credible interval binary presence/absence
\item \verb{SDM_bin_upper_\{sppselect\}}: Upper 95\% credible interval binary presence/absence
\item \verb{SDM_cred_int_width_\{sppselect\}}: Credible interval probability width (i.e., Upper CI - Lower CI)

}
}
\description{
This function writes GeoTIFFs of BART predictions.
}
\details{
The function write outputs of both binary and probability surfaces at mean values and 95\% credible intervals.
}
\note{
Undeclared variables: \code{BART_dir}, \code{sppselect} and \code{tss_threshold}
}
\examples{
\dontrun{
write_bart_rasters(map)
}
}
\author{
Dominic Henry
}
