% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_occ_temporal.R
\name{plot_occ_temporal}
\alias{plot_occ_temporal}
\title{Create temporal plots of species occurrence data}
\usage{
plot_occ_temporal(data, ptheme)
}
\arguments{
\item{data}{\code{sf} point object with species occurrence data}

\item{ptheme}{ggplot theme}
}
\value{
Temporal JPEG plots: 1) \verb{occ_frequency_date_1980_\{sppselect\}} and 2) \verb{occ_frequency_date_all_\{sppselect\}}.
}
\description{
This function summarises occurrence data in a temporal context.
}
\details{
The function draws two plots showing occurrence record frequencies and writes them to the species directory.
}
\note{
The size of the text and other elements of the plots can be change by customising the \code{theme} object. Also take note of the undeclared variables: \code{sdm_dir}, \code{sppselect}, \code{year_date}, and \code{undated}.
}
\examples{
\dontrun{
plot_occ_temporal(occ_data_sf, ptheme)
}
}
\author{
Dominic Henry
}
